Added configurable channel name
Removed the hard coded channel name in the handler. Added a channel
name field to the initial setup form. The bot is fully configurable
to use any username and listen to any channel.

The database schema is capable of supporting multiple bots at the
same time, but the rest of the code only uses the bot with ID=1.